Add individual test reports and errors
This adds individual test and reporting functionality similar to the
given output of pytest-html. As there is no support for collapsing
columns together in Markdown implementations we drop down to bare HTML
so we can use `colspan=3` and make the inline table output for test
results a lot cleaner at the expense of .md readability.

https://github.com/hackebrot/pytest-md/issues/7